requirements.txt
attrs==19.1.0
pendulum==2.0.5
pytest==5.0.1
pyyaml==5.1.1
